:recycle: Move logic for starting ProcessInstances to the adapter This was done to help reduce the complexity of the service.
